Modify core mark

+++ lib/bsp/entry_user.c
@@ -31,7 +31,7 @@ volatile char * const ram = (volatile char*)RAM_BASE_ADDR;
 extern char _heap_start[];
 extern char _heap_end[];
 
-extern volatile unsigned int g_wake_up[2];
+extern volatile uint64_t g_wake_up[2];
 
 void thread_entry(int core_id)
 {
